src/core/privacy_enhancement.py
# E:\AI_Nirvana-1\src\core\privacy_enhancement.py

import numpy as np
from cryptography.fernet import Fernet
from scipy.stats import truncnorm
import hashlib
from sklearn.decomposition import PCA

class PrivacyEnhancement:
    def __init__(self, epsilon=1.0):
        self.epsilon = epsilon
        self.key = Fernet.generate_key()
        self.fernet = Fernet(self.key)

    def add_laplace_noise(self, value):
        beta = 1 / self.epsilon
        noise = np.random.laplace(0, beta)
        return value + noise
    
    def apply_differential_privacy(self, data):
        """对数值数据列表应用差分隐私。"""
        return [self.add_laplace_noise(value) for value in data]

    def set_privacy_budget(self, epsilon):
        """设置隐私预算(epsilon)。"""
        self.epsilon = epsilon

    def encrypt_data(self, data):
        """使用对称加密加密数据。"""
        return self.fernet.encrypt(str(data).encode())

    def decrypt_data(self, encrypted_data):
        """解密使用对称加密的数据。"""
        return self.fernet.decrypt(encrypted_data).decode()

    def k_anonymity(self, data, k):
        """
        对数据应用k-匿名化。
        这是一个简化的实现,仅用于演示目的。
        """
        sorted_data = sorted(data)
        anonymized_data = []
        for i in range(0, len(sorted_data), k):
            group = sorted_data[i:i+k]
            anonymized_value = sum(group) / len(group)
            anonymized_data.extend([anonymized_value] * len(group))
        return anonymized_data

    def gaussian_noise(self, data, mean=0, std=1):
        """添加高斯噪声到数据。"""
        return [value + np.random.normal(mean, std) for value in data]

    def truncated_gaussian_noise(self, data, mean=0, std=1, lower_bound=-1, upper_bound=1):
        """添加截断高斯噪声到数据。"""
        a, b = (lower_bound - mean) / std, (upper_bound - mean) / std
        return [value + truncnorm.rvs(a, b, loc=mean, scale=std) for value in data]

    def hash_data(self, data):
        """对数据进行哈希处理。"""
        return [hashlib.sha256(str(value).encode()).hexdigest() for value in data]

    def dimensionality_reduction(self, data, n_components=2):
        """使用PCA进行降维。"""
        pca = PCA(n_components=n_components)
        return pca.fit_transform(data)

    def randomized_response(self, data, p=0.75):
        """
        应用随机响应技术。
        p是回答真实答案的概率。
        """
        return [value if np.random.random() < p else not value for value in data]

    def exponential_mechanism(self, data, utility_function, sensitivity):
        """
        应用指数机制。
        utility_function应该是一个接受数据项并返回效用分数的函数。
        """
        scores = [utility_function(item) for item in data]
        probabilities = np.exp(self.epsilon * np.array(scores) / (2 * sensitivity))
        probabilities /= probabilities.sum()
        return np.random.choice(data, p=probabilities)

    def local_differential_privacy(self, data, epsilon=1.0):
        """
        应用局部差分隐私。
        这是一个简化的实现,使用随机响应。
        """
        p = np.exp(epsilon) / (1 + np.exp(epsilon))
        return [1 if (value == 1 and np.random.random() < p) or (value == 0 and np.random.random() > p) else 0 for value in data]

    def secure_aggregation(self, data_chunks):
        """
        实现安全聚合。
        这是一个简化的实现,假设所有参与者都是诚实的。
        """
        return sum(sum(chunk) for chunk in data_chunks)

class AdvancedAnonymization:
    def t_closeness(self, data, sensitive_attribute, t):
        """
        应用t-接近性。
        这是一个简化的实现。
        """
        # 实现t-接近性的逻辑
        pass

    def l_diversity(self, data, sensitive_attribute, l):
        """
        应用l-多样性。
        这是一个简化的实现。
        """
        # 实现l-多样性的逻辑
        pass

class FederatedLearningSimulator:
    def simulate_federated_learning(self, client_data, aggregation_function):
        """
        模拟联邦学习过程。
        """
        # 模拟联邦学习的逻辑
        pass

class HomomorphicEncryptionSimulator:
    def simulate_homomorphic_encryption(self, data):
        """
        模拟同态加密操作。
        这是一个简化的实现,仅用于演示目的。
        """
        # 模拟同态加密的逻辑
        pass